code improvements
added one more denying message in bot_requests violation

+++ giorgobot.py
@@ -43,8 +43,6 @@ def identify_member_position(member):
     
     return 0
 
-     
-            
 
 #Το μέρος, όπου οι χρήστες παίρνουν ρόλο βάσει των reactions τους.
 @client.event
@@ -90,7 +88,6 @@ async def on_raw_reaction_add(payload):
 
     #ύστερα, δίνουμε τον ρόλο σε αυτόν που έκανε το react με την φτιαχτή συνάρτησή μας
     await give_role(reactor, role)
-
 
 @client.event
 async def on_raw_reaction_remove(payload):
@@ -138,7 +135,6 @@ async def on_raw_reaction_remove(payload):
     #ύστερα, βγάζουμε τον ρόλο σε αυτόν που έκανε το react με την φτιαχτή συνάρτησή μας
     await remove_role(reactor, role)
 
-
 @client.event
 async def on_message(message):
     #log του μηνύματος.
@@ -158,8 +154,6 @@ async def on_message(message):
 
     #οι ρόλοι που αναπαρίστανται στον σέρβερ
     metzi_tou_neoukti = server.get_role(488730147894198273)
-    pcmci             = server.get_role(488730461091135488)
-    me_meson          = server.get_role(654344275412385793)
 
     #οι admin
     GeorgeMC2610      = client.get_user(250721113729007617)
@@ -167,13 +161,13 @@ async def on_message(message):
 
     #Μετατρέπουμε κάθε μήνυμα σε πεζά γράμματα.
     message.content = message.content.lower()
-    respondable_messages = ["!ping", "!help", "-p", "-play", "-s", "-skip", "-ping", "-leave", "-l", "-help"]
+    respondable_messages = ["!ping", "!help", "-", "!"]
     admin_commands = ["!display members", "!prune"]
 
-    #                                                                   Εκτέλεση εντολών διαχειριστών
-    if (message.content not in respondable_messages and message.content in admin_commands) or ([i for i in admin_commands if message.content.startswith(i)] != []):
+    #Εκτέλεση εντολών διαχειριστών
+    if [i for i in admin_commands if message.content.startswith(i)] != []:
         #Ελέγχουμε αν όντως ο διαχειριστής εκτελεί εντολές.
-        if message.author != Sotiris168 and message.author != GeorgeMC2610:
+        if identify_member_position(message.author) != 4:
             msg_to_send = "Καλή προσπάθεια, " + message.author.mention + "! Αυτή είναι εντολή διαχειριστή. Θα 'ταν κρίμα αν το μάθαιναν οι " + metzi_tou_neoukti.mention + "..."
             await message.channel.send(msg_to_send)
             return
@@ -202,21 +196,29 @@ async def on_message(message):
                 times = int(message_content_by_space[1])
 
                 #δεν πρέπει να 'ναι παραπάνω από πενήντα τα μηνύματα που θα σβησθούν.
-                if times > 50 or times < 0:
+                if times > 50:
                     await message.channel.send("Τι λέτε, κύριε; ΜΑΞ ΠΕΝΗΝΤΑ MHNYMATA, ΚΑΙ ΠΟΛΛΕΣ ΕΙΝΑΙ.")
+                    return
+                elif times < 0:
+                    await message.channel.send("Και για πες, ρε βλάκα, ΠΩΣ ΘΑ ΣΒΗΣΩ **ΑΡΝΗΤΙΚΟ** ΑΡΙΘΜΟ ΜΗΝΥΜΑΤΩΝ;")
                     return
 
                 #αλλιώς, δεν υπάρχει κανένα πρόβλημα και σβήνουμε τα μηνύματα.
+                await message.delete()
                 async for message_to_be_deleted in message.channel.history(limit=times):
                     await message_to_be_deleted.delete()
                 return
             except:
-                await message.channel.send("Ε, καλά, είσαι και πολύ **μαλάκας**. ΑΡΙΘΜΟ ΔΩΣΕ, ΡΕ ΠΟΥΣΤΑΡΕ. \n\n`σωστός χειρισμός: !prune <αριθμός μηνυμάτων για σβήσιμο>`")
+                await message.channel.send("Ε, καλά, είσαι και πολύ **μαλάκας**. ΑΡΙΘΜΟ ΔΩΣΕ, ΡΕ ΠΟΥΣΤΑΡΕ. \n\n**σωστός χειρισμός:** `!prune <αριθμός μηνυμάτων (από 1-50) για σβήσιμο>`")
                 return
                 
-    #                                                                  Εκτέλεση εντολών κοινής χρήσης
-    if message.content in respondable_messages and message.content not in admin_commands:
+    #Εκτέλεση εντολών κοινής χρήσης
+    if [i for i in respondable_messages if message.content.startswith(i)] != []:
 
+        #αν κάποιος χρήστης έχει στείλει απλά μια πάυλα στην αρχή, τότε δεν χρειάζεται να κάνουμε κάτι
+        if message.content[0] == "-" and message.content[-1] == "-":
+            return
+        
         #Στην αρχή βλέπουμε αν το μήνυμα που εστάλη είναι στα bot requests. Αν δεν είναι, δεν εκτελείται η εντολή, σβήνεται η εντολή που εστάλη παράλληλα με το μήνυμα της ειδοποίησης με παράταση 5 δευτερολέπτων.
         if message.channel != bot_requests:
 
@@ -226,18 +228,19 @@ async def on_message(message):
             deny3 = "Κάθε φορά που στέλενεις εντολή έξω από το" + bot_requests.mention + " ένα κουταβάκι πεθαίνει... 😥"
             deny4 = "Γράψε 100 φορές στο τετράδιο σου 'ΘΑ ΣΤΕΛΝΩ ΤΙΣ ΕΝΤΟΛΕΣ ΜΟΥ ΜΟΝΟ ΣΤΟ " + bot_requests.mention + "'." 
             deny5 = "Στείλ' το στο " + bot_requests.mention + ", αλλιώς θα το πω στην κυρίααα 😨."
-            deny6 = "🤡  <-- εσύ, όταν δεν στέλενεις τις εντολές σου στο " + bot_requests.mention + "."
+            deny6 = "🤡  ← εσύ, όταν δεν στέλενεις τις εντολές σου στο " + bot_requests.mention + "."
             deny7 = "Θα έβαζες ποτέ το ψυγείο στο μπαλκόνι; Όχι. Μην βάζεις εντολές **έξω** του " + bot_requests.mention + ", τότε **__ΒΛΑΚΑ__**."
             deny8 = "Έχω πει 500 135.000 φορές να τα στέλνεις στο " + bot_requests.mention + "..."
+            deny9 = "🚓🚓 **ΑΣΤΥΝΟΜΙΑ ΒΛΑΚΕΙΑΣ!** Ήμουν σίγουρος, ότι κάποιος σαν κι εσένα, θα έστελνε εντολή εκτός του " + bot_requests.mention + "!"
 
-            denying_messages = [deny1, deny2, deny3, deny4, deny5, deny6, deny7, deny8]
+            denying_messages = [deny1, deny2, deny3, deny4, deny5, deny6, deny7, deny8, deny9]
 
             #επιλέγουμε ένα τυχαίο από αυτά
-            random_selection = random.randint(0, len(denying_messages)-1)
+            random_deny_message = random.choice(denying_messages)
 
             #σβήνουμε το μήνυμα του χρήστη, και μετά αυτό που στέλενει το bot
             await message.delete()
-            await message.channel.send(denying_messages[random_selection], delete_after=8.0) 
+            await message.channel.send(random_deny_message, delete_after=8.0) 
             return
         
         #Εκτέλεση των εντολών
@@ -246,23 +249,22 @@ async def on_message(message):
             return
 
         if message.content == respondable_messages[1]:
-            help_message = "**ΔΙΚΕΣ ΜΟΥ ΕΝΤΟΛΕΣ:** \n `!help` --> Δείχνει το παρόν μενού.\n `!ping` --> ανταπόκριση του μποτ με 'Pong!'.\n\n **ΕΝΤΟΛΕΣ ΔΙΑΧΕΙΡΙΣΤΗ:**\n `!display users` --> Προβολή όλων των μελών του σέρβερ.\n `!prune <αριθμός 1-50>` --> Σβήσιμο όλων των προηγούμενων μηνυμάτων\n `!secret santa` --> Νέα κλήρωση για secret santa."
+            help_message = "**ΔΙΚΕΣ ΜΟΥ ΕΝΤΟΛΕΣ:** \n `!help` --> Δείχνει το παρόν μενού.\n `!ping` --> ανταπόκριση του μποτ με 'Pong!'.\n\n **ΕΝΤΟΛΕΣ ΔΙΑΧΕΙΡΙΣΤΗ:**\n `!display users` --> Προβολή όλων των μελών του σέρβερ.\n `!prune <αριθμός 1-50>` --> Σβήσιμο όλων των προηγούμενων μηνυμάτων"
             await message.channel.send(help_message)
             return
             
-    #                                               Εδώ ελέγχουμε αν έχει σταλεί κάποιο μήνυμα σε library χωρίς φωτογραφία
-    if message.channel.category_id == 749958245203836939:
-        if not message.attachments:
-            deny1 = "Σσσσσσσσσσσσσσσσσσσσσς! ***ΨΥΘΙΡΣΤΑ*** εδώ είναι βιβλιοθήκη! Δεν κάνει να μιλάμε εδώ..."
-            deny2 = "Ρε κλόουν. Όχι μηνύματα εδώ. ΜΟΝΟ ΦΩΤΟΓΡΑΦΙΕΣ/ΒΙΝΤΕΟ."
-            deny3 = "🚓 ΣΥΛΛΑΜΒΑΝΕΣΑΙ, ΒΛΑΚΑΚΟ. ΜΙΛΟΥΣΕΣ ΣΤΗ ΒΙΒΛΙΟΘΗΚΗ. 10 μέρες φυλακή μέχρι να μάθεις να στέλνεις μόνο φωτογραφίες ή βίντεο."
-            deny4 = "Εδώ. Φωτογραφίες/Βίντεο. ***__ΜΟΝΟ__***. Εχμ, ωραία :)"
+    #Εδώ ελέγχουμε αν έχει σταλεί κάποιο μήνυμα σε library χωρίς φωτογραφία
+    if message.channel.category_id == 749958245203836939 and not message.attachments:
+        deny1 = "Σσσσσσσσσσσσσσσσσσσσσς! ***ΨΥΘΙΡΣΤΑ*** εδώ είναι βιβλιοθήκη! Δεν κάνει να μιλάμε εδώ..."
+        deny2 = "Ρε κλόουν. Όχι μηνύματα εδώ. ΜΟΝΟ ΦΩΤΟΓΡΑΦΙΕΣ/ΒΙΝΤΕΟ."
+        deny3 = "🚓 ΣΥΛΛΑΜΒΑΝΕΣΑΙ, ΒΛΑΚΑΚΟ. ΜΙΛΟΥΣΕΣ ΣΤΗ ΒΙΒΛΙΟΘΗΚΗ. 10 μέρες φυλακή μέχρι να μάθεις να στέλνεις μόνο φωτογραφίες ή βίντεο."
+        deny4 = "Εδώ. Φωτογραφίες/Βίντεο. ***__ΜΟΝΟ__***. Εχμ, ωραία :)"
 
-            denying_messages = [deny1, deny2, deny3, deny4]
+        denying_messages = [deny1, deny2, deny3, deny4]
 
-            random_selection = random.randint(0, len(denying_messages)-1)
-            await message.delete()
-            await message.channel.send(denying_messages[random_selection], delete_after=8.0)
-            return
+        random_deny_message = random.choice(denying_messages)
+        await message.delete()
+        await message.channel.send(random_deny_message, delete_after=8.0)
+        return
 
 client.run(token)
